dbmesh/pyproject.toml
[project]
name = "dbmesh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.6.0",  # MCP server
    "fastapi>=0.115.4",  # FastAPI framework
    "uvicorn>=0.31.0",  # ASGI server
    "pydantic>=2.10.2",  # Data validation
    "psycopg2-binary>=2.9.10",  # PostgreSQL adapter
    "sqlalchemy>=2.0.31",  # ORM
    "pyyaml>=6.0",  # YAML config
]
